requirements.txt
certifi==2020.6.20
click==7.1.2
dill==0.3.2
numpy==1.19.2
pandas==1.1.2
pyfunctional==1.4.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
tabulate==0.8.7
